Create parser and split classes from "Command"
Создал класс, который парсит введённые данные. Разнёс классы в отдельные файлы. Теперь в классе распределения должен быть метод, который проверяет аргументы на правильность.
