+++ BotUser.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Text.Json.Serialization;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 
 namespace kontur_project
 {
@@ -11,14 +6,14 @@ namespace kontur_project
     {
         public long UserId { get; set; }
         public Stack<ICondition> UserConditions { get; set; }
-        public List<IDistribution> Distribution { get; set; }
+        public List<Distribution> Distribution { get; set; }
         public List<string> Methods { get; set; }
         public List<double> Args { get; set; }
 
         public BotUser(long userId)
         {
             UserId = userId;
-            Distribution = new List<IDistribution>();
+            Distribution = new List<Distribution>();
             Methods = new List<string>();
             Args = new List<double>();
             UserConditions = new Stack<ICondition>();
