+++ UniformDistribution.cs
@@ -1,0 +1,41 @@
+﻿
+namespace kontur_project
+{
+    class UniformDistribution : Distribution
+    {
+        public override long Id { get; set; }
+        public override int ParamNum { get => 2; }
+        public override string Name => "Равномерное";
+        public override double[] distParams { get; set; }
+
+        public override double Density(double x)
+        {
+            if (x <= distParams[0])
+                return 0;
+            if (distParams[0] < x && x < distParams[1])
+                return 1 / (distParams[1] - distParams[0]);
+            else
+                return 0;
+        }
+
+        public override double ProbabilityFunction(double x)
+        {
+            if (x <= distParams[0])
+                return 0;
+            if (distParams[0] < x && x < distParams[1])
+                return (x - distParams[0]) / (distParams[1] - distParams[0]);
+            else
+                return 1;
+        }
+
+        public override string CheckParamsValid(params double[] parametres)
+        {
+            var answer = BaseParamsCheking(parametres);
+
+            if (parametres[0] > parametres[1])
+                answer = "Параметры равномерного распределения должны задавать отрезок неотрицательной длины";
+
+            return answer;
+        }
+    }
+}
